New score function

+++ HashCode2017/utils.h
@@ -22,7 +22,7 @@
 
 void printSolution(FILE* file, vector<vector<int>>& result);
 void checkSolution(vector<vector<int>>& result, vector<int> sizesVideo, int maxSize);
-int getScore(vector<vector<int>>& result, vector<struct point>& endPoints);
+int64_t getScore(vector<vector<int>>& result, vector<struct point>& endPoints);
